requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
aiosqlite==0.20.0
apscheduler==3.10.4
google-cloud-bigquery==3.25.0
google-auth==2.35.0
pydantic-settings==2.5.2
httpx==0.27.2
pytz==2024.2
